inference_evaluation/evaluation/requirements.txt
psycopg2-binary>=2.9.0
func-timeout>=4.3.5
numpy>=1.24.0
